Fix file rename without move
In the case of a file being renamed within a watched folder we get the
following sequence of events from notify:
- Modify(Name(From))
- Modify(Name(To))
- Modify(Name(Both))

The recent update to treat `Modify(Name(From))` as a remove (to
correctly handle files moved out of tracked folders entirely) means that
the renamed file was first getting marked as removed (incorrect) and
then getting renamed (correct). More sophisticated deduplication of
events is probably the *right* solution, but simply ensuring that the
file is not marked as removed when handling a rename solves the
immediate problem.